+++ data/duke.txt
@@ -1,3 +1,3 @@
 1.[T][X] a
-2.[D][ ] b       |by: 2012-12-12T12:21:00
-3.[E][X] c       |from: 12pm to: 1pm
+2.[D][ ] b        |by: 2012-12-12T12:21:00
+3.[E][X] c        |from: 12pm to: 1pm
