+++ src/main/java/duke_exception/DukeException.java
@@ -1,4 +1,4 @@
-package DukeException;
+package duke_exception;
 class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
