+++ src/main/java/task/ToDo.java
@@ -1,4 +1,4 @@
-package Task;
+package task;
 public class ToDo extends Task {
 
     public ToDo(String description) {
