+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hello! I'm Duke
+Hello! I'm duke
 What can I do for you?
 Got it. I've added this task:
   [T][ ]  borrow book
